Reorganized Indexer files.
