+++ src/utilities/indexer/index-parsers.ts
@@ -1,6 +1,6 @@
 import { ethers } from "ethers";
 
-import { Indexer } from "../types";
+import { Indexer } from "../../types";
 
 //=======================================
 // NounsDAO
